Finishes prototype of services.

+++ TransmitterIndex/transmitter_index.py
@@ -2,11 +2,12 @@ import flask
 import sys
 import os
 import json
-import threading
+import requests
 import time
+from jinja2 import Template
 
-from transmitter import transmitter_thread
-from scanner import scan
+from transmitter import run_transmitter
+from scanner import run_scanner
 
 if len(sys.argv) < 2:
 
@@ -14,37 +15,53 @@ if len(sys.argv) < 2:
 
 else:
 
-    port = int(sys.argv[1])
+    port = sys.argv[1]
+    run_transmitter(port, 10)
+    port = int(port)
 
-    transmitter_thread('Transmitter Index', 'A list of services the space.', port)
-    
-    def scanner_thread():
-        
-        while True:
-            known_transmitters = scan()
-            known_services = []
-            for transmitter in known_transmitters:
-              ip = transmitter[0]
-              services = transmitter[1]
-              for service in services:
-                known_services += [{
-                  'name': service['name'],
-                  'desc': service['desc'],
-                  'url': 'http://%s:%i' % (ip, service['port'])
-                }]
-            
-            with open('services.json', 'w') as f:
-                f.write(json.dumps(known_services))
-            
-            time.sleep(60)
-    
-    threading.Thread(target = scanner_thread).start()
-    
-    
+    def scanner_thread(addr):
+
+        now = time.time()
+        service = {
+            'url': 'http://%s:%s' % (addr[0], addr[1]), 'last_seen': now}
+        try:
+            r = requests.get(service['url'] + '/transmitter_info')
+            if r.status_code == 200:
+                info = json.loads(r.text)
+                service['name'] = info['name']
+                service['description'] = info['description']
+
+                if not os.path.isfile('services.json'):
+                    with open('services.json', 'w') as f:
+                        f.write(json.dumps([]))
+
+                with open('services.json', 'r') as f:
+                    known_services = json.loads(f.read())
+
+                known_services = [
+                    old_service for old_service in known_services
+                    if old_service['url'] != service['url'] and 30 > now - old_service['last_seen']
+                ]
+
+                if service not in known_services:
+                    with open('services.json', 'w') as f:
+                        f.write(json.dumps(known_services +
+                                           [service]))
+        except requests.exceptions.ConnectionError:
+            pass
+
+    run_scanner(scanner_thread)
 
     app = flask.Flask(__name__)
 
-    @app.route('/', methods = ['GET'])
+    @app.route('/transmitter_info', methods=['GET'])
+    def get_transmitter_info():
+        return json.dumps({
+            'name': 'Transmitter Index',
+            'description': 'A list of transmitters currently in the space.'
+        })
+
+    @app.route('/', methods=['GET'])
     def get_services():
 
         if os.path.isfile('services.json'):
@@ -52,24 +69,15 @@ else:
                 services = json.loads(f.read())
         else:
             services = []
-        
-        html = '''
-<html>
-  <body>
-    <dl>'''
-      
-        if 0 == len(services):
-            html += '      <dt>No services.</dt>'
-        else:
-            for service in services:
-                html += '      <dt><a href="%s">%s</a></dt>' % (service['url'],service['name'])
-                html += '      <dd>%s</dd>' % service['desc']
-      
-        html += '''
-    </dl>
-  </body>
-</html>'''
-      
+
+        services.sort(key=lambda x: x['name'])
+
+        with open('transmitter_index.html', 'r') as f:
+            html = Template(f.read()).render({
+                'stylesheet': flask.url_for('static', filename='pink_on_black.css'),
+                'transmitters': services
+            })
+
         return html
 
-    app.run(host = '0.0.0.0', port = port)
+    app.run(host='0.0.0.0', port=port)
